Refactor using @amrc-factoryplus/edge-driver
