+++ edge-test/Makefile
@@ -1,0 +1,6 @@
+top=..
+include ${top}/mk/acs.init.mk
+
+repo?=edge-test
+
+include ${mk}/acs.js.mk
